Split resource steps
