+++ frictionless/steps/resource/__init__.py
@@ -1,0 +1,4 @@
+from .resource_add import resource_add
+from .resource_remove import resource_remove
+from .resource_transform import resource_transform
+from .resource_update import resource_update
